+++ Assignment.sln
@@ -8,9 +8,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		Assignment\Assignment\Directory.Build.props = Assignment\Assignment\Directory.Build.props
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Assignment", "Assignment\Assignment\Assignment.csproj", "{0030CB91-2D00-402A-ABBA-1235F27FBE38}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Assignment.Tests", "Assignment.Tests\Assignment.Tests.csproj", "{C1126C49-15F6-49D4-981C-520FD7080608}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Assignment", "Assignment\Assignment.csproj", "{118878BC-0576-4FFE-867A-BF747267D8C5}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -18,14 +18,14 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{0030CB91-2D00-402A-ABBA-1235F27FBE38}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0030CB91-2D00-402A-ABBA-1235F27FBE38}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0030CB91-2D00-402A-ABBA-1235F27FBE38}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0030CB91-2D00-402A-ABBA-1235F27FBE38}.Release|Any CPU.Build.0 = Release|Any CPU
 		{C1126C49-15F6-49D4-981C-520FD7080608}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C1126C49-15F6-49D4-981C-520FD7080608}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C1126C49-15F6-49D4-981C-520FD7080608}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C1126C49-15F6-49D4-981C-520FD7080608}.Release|Any CPU.Build.0 = Release|Any CPU
+		{118878BC-0576-4FFE-867A-BF747267D8C5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{118878BC-0576-4FFE-867A-BF747267D8C5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{118878BC-0576-4FFE-867A-BF747267D8C5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{118878BC-0576-4FFE-867A-BF747267D8C5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
